requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.8.6
coverage==7.2.1